web/requirements-dev.txt
-r requirements.txt
# Debugging
# -------------------------------------
django-debug-toolbar==1.4
ipdb==0.8.1
ipython==4.0.1

# Testing and coverage
# -------------------------------------
pytest==3.0.3
pytest-django==2.9.1
pytest-pythonpath==0.7.1
pytest-cov==2.3.1
pdbpp==0.8.3
flake8==3.0.4
